Pin brace-expansion >=5.0.6 and extract demo-shared helpers
OSV-Scanner flagged CVE-2026-45149 (GHSA-jxxr-4gwj-5jf2) on
brace-expansion@5.0.5, fixed in 5.0.6. Pinned via a pnpm override
scoped to the 5.x line so minimatch@9.0.9's brace-expansion@2.x dep
is unaffected.

SonarCloud's quality gate failed on new_duplicated_lines_density
(32.7% > 3% threshold) because the three demo scenes each carried
the same defensive ctx narrowing, the same setActive helper, and
the same scene-root query. Extracted them into
src/scenes/demo-shared.ts so per-scene files now hold only their
scene-specific constants and lifecycle hooks. Per-scene constants
(beat names, copy text, asset paths) stay in the scene files per
the architecture preflight's "no demo framework on first pass".
