+++ package.json
@@ -36,5 +36,10 @@
   "dependencies": {
     "gsap": "^3.15.0",
     "howler": "^2.2.4"
+  },
+  "pnpm": {
+    "overrides": {
+      "brace-expansion@^5": ">=5.0.6"
+    }
   }
 }
