+++ src/scenes/demo-feature.ts
@@ -10,24 +10,25 @@
 // require). The Playwright spec at
 // `tests-e2e/demo-composition.spec.ts` HEAD-fetches the asset URL so
 // "the committed path is actually served" is a runtime check, not just
-// a static-validation pass.
+// a static-validation pass. Shared defensive helpers (ctx predicate,
+// activation flip, scene-root query) live in `./demo-shared.ts`.
 
-import { NAVIGATION_MODES } from '../runtime/navigation';
 import type { SceneModule } from '../runtime/scene';
-import type { WorkbenchSceneCtx } from '../runtime/scene-loader';
+import {
+  DEMO_ACTIVE_ATTR,
+  DEMO_ROOT_ATTR,
+  findDemoRoot,
+  isDemoCtx,
+  setDemoActive,
+} from './demo-shared';
 
 // Exported so the unit + e2e tests assert against ONE constant rather
 // than each duplicating the literal path. A future asset replacement
 // is a one-line edit here.
 export const DEMO_FEATURE_ASSET_URL = '/assets/demo/pulsar-mark.svg';
 
-const ROOT_ATTR = 'data-pulsar-demo-scene';
 const ROOT_VALUE = 'demo-feature';
 const STATE_ATTR = 'data-pulsar-demo-state';
-// See `demo-title.ts` for the rationale on the activation marker —
-// timeline-owned activation keeps the demo composition sequential
-// even though the resolver mounts every scene's root up front.
-const ACTIVE_ATTR = 'data-pulsar-demo-active';
 const FEATURE_IMG_ATTR = 'data-pulsar-demo-feature-img';
 
 const BEAT_FEATURE_REVEAL = 'feature-reveal';
@@ -35,79 +36,6 @@ const BEAT_FEATURE_REVEAL = 'feature-reveal';
 const HEADLINE_TEXT = 'Composition over slides.';
 const BODY_TEXT =
   'Scenes carry their own timelines, captions, assets, and audio. Compositions arrange them.';
-
-interface FixtureDomElement {
-  setAttribute(name: string, value: string): void;
-  appendChild?(node: unknown): unknown;
-  textContent?: string;
-}
-
-interface FixtureDomFactory {
-  createElement(tag: string): FixtureDomElement;
-}
-
-interface FixtureStageElement {
-  setAttribute(name: string, value: string): void;
-  removeAttribute(name: string): void;
-  appendChild?(node: unknown): unknown;
-  querySelector?(selector: string): {
-    setAttribute?(name: string, value: string): void;
-    remove?(): void;
-  } | null;
-  readonly ownerDocument?: FixtureDomFactory | null;
-}
-
-interface DemoCtx {
-  readonly stage: FixtureStageElement | null;
-  readonly mode: WorkbenchSceneCtx['mode'];
-  readonly gsap: WorkbenchSceneCtx['gsap'];
-}
-
-const isStageShape = (stage: unknown): stage is FixtureStageElement | null => {
-  if (stage === null) return true;
-  if (typeof stage !== 'object') return false;
-  const candidate = stage as Partial<Record<'setAttribute' | 'removeAttribute', unknown>>;
-  return (
-    typeof candidate.setAttribute === 'function' && typeof candidate.removeAttribute === 'function'
-  );
-};
-
-const isGsapShape = (gsap: unknown): gsap is WorkbenchSceneCtx['gsap'] => {
-  if (gsap === null || typeof gsap !== 'object') return false;
-  return typeof (gsap as { timeline?: unknown }).timeline === 'function';
-};
-
-const isDemoCtx = (value: unknown): value is DemoCtx => {
-  if (typeof value !== 'object' || value === null) return false;
-  if (!('stage' in value) || !('mode' in value) || !('gsap' in value)) return false;
-  const { stage, mode, gsap } = value;
-  if (!isStageShape(stage)) return false;
-  if (typeof mode !== 'string' || !(NAVIGATION_MODES as readonly string[]).includes(mode)) {
-    return false;
-  }
-  return isGsapShape(gsap);
-};
-
-const findRoot = (
-  stage: FixtureStageElement,
-): {
-  setAttribute?(name: string, value: string): void;
-  remove?(): void;
-} | null => {
-  if (typeof stage.querySelector !== 'function') return null;
-  return stage.querySelector(`[${ROOT_ATTR}="${ROOT_VALUE}"]`);
-};
-
-const setActive = (
-  root: {
-    setAttribute?(name: string, value: string): void;
-  } | null,
-  active: boolean,
-): void => {
-  if (root === null || typeof root.setAttribute !== 'function') return;
-  root.setAttribute(ACTIVE_ATTR, active ? 'true' : 'false');
-  root.setAttribute('style', active ? '' : 'display: none;');
-};
 
 export const demoFeatureScene: SceneModule = {
   id: 'demo-feature',
@@ -135,10 +63,9 @@ export const demoFeatureScene: SceneModule = {
     if (typeof ownerDoc.createElement !== 'function') return;
 
     const root = ownerDoc.createElement('section');
-    root.setAttribute(ROOT_ATTR, ROOT_VALUE);
+    root.setAttribute(DEMO_ROOT_ATTR, ROOT_VALUE);
     root.setAttribute(STATE_ATTR, 'mounted');
-    // Initially inactive — see demo-title.ts for the rationale.
-    root.setAttribute(ACTIVE_ATTR, 'false');
+    root.setAttribute(DEMO_ACTIVE_ATTR, 'false');
     root.setAttribute('style', 'display: none;');
     if (typeof root.appendChild === 'function') {
       const heading = ownerDoc.createElement('h2');
@@ -165,14 +92,11 @@ export const demoFeatureScene: SceneModule = {
     if (!isDemoCtx(ctx)) return null;
     const stage = ctx.stage;
     if (stage === null) return null;
-    const root = findRoot(stage);
+    const root = findDemoRoot(stage, ROOT_VALUE);
     const tl = ctx.gsap.timeline();
-    tl.call(() => setActive(root, true));
+    tl.call(() => setDemoActive(root, true));
     tl.to({}, { duration: 1 });
     tl.addLabel(BEAT_FEATURE_REVEAL, 1);
-    // Trailing mutation rides on the last tween's `onComplete` — see
-    // `src/scenes/demo-title.ts` for the rationale (last-entry master
-    // boundary race).
     tl.to(
       {},
       {
@@ -181,7 +105,7 @@ export const demoFeatureScene: SceneModule = {
           if (root !== null && typeof root.setAttribute === 'function') {
             root.setAttribute(STATE_ATTR, 'ran');
           }
-          setActive(root, false);
+          setDemoActive(root, false);
         },
       },
     );
@@ -191,7 +115,7 @@ export const demoFeatureScene: SceneModule = {
     if (!isDemoCtx(ctx)) return;
     const stage = ctx.stage;
     if (stage === null) return;
-    const root = findRoot(stage);
+    const root = findDemoRoot(stage, ROOT_VALUE);
     if (root !== null && typeof root.remove === 'function') {
       root.remove();
     }
